Style: 코드리뷰 반영 state 제거, Star 컴포넌트 합치기
